+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-# v0.1.0
-##  01/06/2017
+# v1.0.0
+##  01/11/2017
 
 1. [](#new)
-    * ChangeLog started...
+    * Simply create QR Codes with the shortcode `[qrcode]`
+    * Override the QR Code parameters for each page
+    * Inline parameters for the QR Code shortcode element    
+    * Button to add Qr Codes in the content editor
+    * Twig functions to embed QR Codes in your templates
+    * Multi-Language Support